Massive marking of strings as "const".
Unfortunately it seems the codebase is not very good about tracking lifetimes
of strings, otherwise this might be easier to pull off.  May objects will just
retain a string pointer and assume it will be around forever.  There are some
places that would be better as std::string.  This is, by comparison, a more
shallow cleanup.