Make frame buttons and icons non-const dimensions; add QvImage scaling.
One of my earliest experiments with this WM ws trying to raise image
dimensions.  I found when I did that by editing fbutton.h and qvwm.h,
the images for icons and minimize, maximize, or close buttons were not
scaled.  This commit introduces that, so I can raise the dimensions on
a high DPI display and it only looks a *little* terrible.

The scaling algorithm used here is quite bad.  Perhaps it would be good
to write XRender code and leave in the current thing [or an improved
version] as a fallback for ancient X servers.